Split IntegrationTest into several smaller traits
