+++ node/src/main/scala/fluence/node/tendermint/json/ValidatorKey.scala
@@ -16,10 +16,13 @@
 
 package fluence.node.tendermint.json
 
+import java.util.Base64
+
 import fluence.ethclient.helpers.Web3jConverters.base64ToBytes32
 import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
 import io.circe.{Decoder, Encoder}
 import org.web3j.abi.datatypes.generated.Bytes32
+import scodec.bits.ByteVector
 
 /**
  * Validator's public key in Tendermint-compatible format.
