+++ node/src/test/scala/fluence/node/GanacheSetup.scala
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2018 Fluence Labs Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fluence.node
+import cats.syntax.functor._
+import org.scalatest.{Timer => _}
+
+import scala.language.higherKinds
+
+trait GanacheSetup extends OsSetup with slogging.LazyLogging {
+
+  def wireupContract(): Unit = {
+    logger.info("bootstrapping npm")
+    runCmd("npm install")
+
+    logger.info("starting Ganache")
+    runBackground("npm run ganache")
+
+    logger.info("deploying contract to Ganache")
+    runCmd("npm run migrate")
+  }
+
+  def killGanache(): Unit = {
+    logger.info("killing ganache")
+    runCmd("pkill -f ganache")
+  }
+}
