+++ node/src/test/scala/fluence/node/TendermintSetup.scala
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2018 Fluence Labs Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fluence.node
+
+import cats.effect._
+import cats.syntax.functor._
+import fluence.node.eth.WorkerNode
+import org.scalatest.{Timer => _}
+
+import scala.io.Source
+import scala.language.higherKinds
+import scala.util.Try
+
+trait TendermintSetup {
+  protected def heightFromTendermintStatus(host: String, p2pPort: Short): IO[Option[Long]] = IO {
+    import io.circe.Json
+    import io.circe.parser.parse
+    val rpcPort = WorkerNode.rpcPort(p2pPort)
+    val source = Source.fromURL(s"http://$host:$rpcPort/status").mkString
+    val height = parse(source)
+      .getOrElse(Json.Null)
+      .asObject
+      .flatMap(_("result"))
+      .flatMap(_.asObject)
+      .flatMap(_("sync_info"))
+      .flatMap(_.asObject)
+      .flatMap(_("latest_block_height"))
+      .flatMap(_.asString)
+      .flatMap(x => Try(x.toLong).toOption)
+    height
+  }
+}
